Change LLR size to 32-bit for Turbo Decoder
Updated the LLR data type from int16_t to int32_t to match
the scalar Turbo Decoder requirements.

Note: While this aligns the LLR size with the decoder,
segmentation faults still occur during UL reception and
further investigation is needed.

+++ openair1/PHY/CODING/3gpplte_turbo_decoder.c
@@ -120,7 +120,8 @@
 
 
 typedef char Binary;
-typedef short llr_t; // internal decoder data is 16-bit fixed
+// Change llr_t from 16-bit to 32-bit to fix decoding issues in the scalar Turbo Decoder
+typedef int32_t llr_t;
 typedef short channel_t;
 
 #define LLR_MAX 32767
